change - separate song and party repository&controller
