+++ ZMA/Controllers/PartyController.cs
@@ -66,72 +66,6 @@ public class PartyController(IPartyRepository partyRepository, UserManager<Host>
         }
     }
 
-    [HttpPost("RequestSong")]
-    public async Task<ActionResult> RequestSong([Required] string title, [Required] Guid partyId)
-    {
-        try
-        {
-            var song = new Song() { Title = title, RequestTime = DateTime.Now, Accepted = false };
-            
-           await partyRepository.RequestSong(song, partyId);
-
-            return Ok("Successfully requested song.");
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e.Message);
-            return NotFound("Requesting song failed.");
-        }
-    }
-
-    [HttpPatch("AcceptSong"), Authorize(Roles = "Host")]
-    public async Task<ActionResult> AcceptSong([Required] int songId)
-    {
-        try
-        {
-            await partyRepository.AcceptSong(songId);
-
-            return Ok("Song accepted by host.");
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e.Message);
-            return NotFound("Accepting song failed.");
-        }
-    }
-
-    [HttpGet("GetSongs"), Authorize(Roles = "Host")]
-    public async Task<ActionResult<ICollection<Song>>> GetSongs([Required] Guid partyId)
-    {
-        try
-        {
-            return Ok(await partyRepository.GetSongs(partyId));
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e.Message);
-            return NotFound("Retrieving songs failed.");
-        }
-    }
-
-    [HttpDelete("DenySong"), Authorize(Roles = "Host")]
-    public async Task<ActionResult> DenySong([Required] Guid partyId, [Required] int songId)
-    {
-        try
-        {
-            await partyRepository.DeleteSong(partyId, songId);
-            
-            logger.LogInfo("Host deleted song.");
-
-            return Ok("Song denied");
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e.Message);
-            return NotFound("Deleting song failed.");
-        }
-    }
-
     [HttpDelete("DeleteParty"), Authorize(Roles = "Host")]
     public async Task<ActionResult> DeleteParty([Required] Guid partyId)
     {
